fix: blobLength type conversions

+++ disperser/apiserver/server_v2_test.go
@@ -621,7 +621,7 @@ func TestInvalidLength(t *testing.T) {
 	require.NoError(t, err)
 
 	// Length we are commiting to should be a power of 2.
-	require.Equal(t, commitments.Length, math.NextPowOf2u64(uint64(commitments.Length)))
+	require.Equal(t, uint64(commitments.Length), math.NextPowOf2u64(uint64(commitments.Length)))
 
 	// Changing the number of commitments should cause an error before a validity check of the commitments
 	commitments.Length += 1
@@ -671,7 +671,7 @@ func TestTooShortCommitment(t *testing.T) {
 	require.NoError(t, err)
 
 	// Length we are commiting to should be a power of 2.
-	require.Equal(t, commitments.Length, math.NextPowOf2u64(uint64(commitments.Length)))
+	require.Equal(t, uint64(commitments.Length), math.NextPowOf2u64(uint64(commitments.Length)))
 
 	// Choose a smaller commitment length than is legal. Make sure it's a power of 2 so that it doesn't
 	// fail prior to the commitment length check.
